Removal of channelMixState object
