+++ RWKV-v5/src/model.py
@@ -3,7 +3,7 @@
 ### ---
 
 from .module.CoreDependencies import *
-from .module.ChannelMix import channelMix_batchForward
+from .module.ChannelMix import channelMix_batchForward, RWKV_ChannelMix
 
 # ---
 # Isolating out known operations that **does not work** with torch.compile
@@ -186,46 +186,6 @@ class RWKV_TimeMix(JITModClass):
         return x_logits, TimeMixState(x[:,-1],ms[-1])
     
 ### ---
-
-
-class RWKV_ChannelMix(JITModClass):
-
-    def __init__(self, layer_id, n_layer, n_embd, dim_ffn):
-        super().__init__()
-        # self.layer_id = layer_id
-        # self.time_shift = nn.ZeroPad2d((0, 0, 1, -1))
-
-        with torch.no_grad():  # fancy init of time_mix
-            ratio_1_to_almost0 = 1.0 - (layer_id / n_layer)  # 1 to ~0
-            ddd = torch.ones(1, 1, n_embd)
-            for i in range(n_embd):
-                ddd[0, 0, i] = i / n_embd
-            self.time_mix_k = nn.Parameter(torch.pow(ddd, ratio_1_to_almost0))
-            self.time_mix_r = nn.Parameter(torch.pow(ddd, ratio_1_to_almost0))
-
-        self.key = nn.Linear(n_embd, dim_ffn, bias=False)
-        self.receptance = nn.Linear(n_embd, n_embd, bias=False)
-        self.value = nn.Linear(dim_ffn, n_embd, bias=False)
-
-    @JITModMethod
-    @TCompileMax
-    def forward(self, x, last_state: torch.Tensor):
-        # out_emb, out_state = channelMix_batchForward(
-        #     self.time_mix_k,self.time_mix_r,
-        #     self.key.bi
-        #     self.key.weight, self.receptance.weight, self.value.weight,
-        #     x, last_state.shift_state
-        # )
-        # return (out_emb, (out_state))
-    
-        xx = torch.concat((last_state.unsqueeze(1), x[:, :-1]),
-                          dim=1)
-        xk = x * self.time_mix_k + xx * (1 - self.time_mix_k)
-        xr = x * self.time_mix_r + xx * (1 - self.time_mix_r)
-        kv = self.value( torch.relu( self.key(xk) ) ** 2 )
-        return (torch.sigmoid(self.receptance(xr)) * kv,
-                (x[:, -1]))
-
 
 ### ---
 # The RWKV Model blocks
